+++ frontend/src/components/FilterSection/SearchSection/index.styled.ts
