+++ frontend/src/components/FilterSection/SearchSection/index.tsx
